feat(ShowBankAccountWebController):created class, and modified show.html accordingly to show my first instance's fields on the web as a String, with float balance field, but not yet with currency set

+++ week-17/day-04/BankOfSimba/src/main/java/com/greenfoxacademy/bankofsimba/Model/BankAccount.java
@@ -3,10 +3,10 @@ package com.greenfoxacademy.bankofsimba.Model;
 public class BankAccount {
 
   private String name;
-  private int balance;
+  private float balance;
   private String animalType;
 
-  public BankAccount(String name, int balance, String animalType) {
+  public BankAccount(String name, float balance, String animalType) {
     this.name = name;
     this.balance = balance;
     this.animalType = animalType;
@@ -20,11 +20,11 @@ public class BankAccount {
     this.name = name;
   }
 
-  public int getBalance() {
+  public double getBalance() {
     return balance;
   }
 
-  public void setBalance(int balance) {
+  public void setBalance(float balance) {
     this.balance = balance;
   }
